requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
Django==5.2.4
django-environ==0.12.0
django_celery_results==2.6.0
djangorestframework==3.16.0
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
kombu==5.5.4
libcst==1.8.2
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
promise==2.3
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.2.0
rich==14.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.47.2
strawberry-graphql==0.278.0
strawberry-graphql-django==0.65.1
text-unidecode==1.3
typer==0.16.0
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.9.0
